Incremental commit - work on DataMgrUtils

+++ include/vapor/DCWRF.h
@@ -95,9 +95,9 @@ public:
  virtual std::vector <string> GetCoordVarNames() const;
 
 
- //! \copydoc DC::GetCoordVarNames()
+ //! \copydoc DC::GetNumRefLevels()
  //!
- virtual int GetNumRefLevels(string varname) const { return(1); }
+ virtual size_t GetNumRefLevels(string varname) const { return(1); }
 
  //! \copydoc DC::GetMapProjection(string)
  //!
